+++ Sources/CubiomesMapKit/Overlays/MinecraftDrawingOverlay.swift
@@ -8,6 +8,22 @@
 import MapKit
 import PencilKit
 
+public struct MinecraftDrawing: MinecraftMapBuilderContent {
+    public var model: MinecraftMapDrawing
+
+    public init(model: MinecraftMapDrawing) {
+        self.model = model
+    }
+
+    public init(drawing: PKDrawing, location: CLLocationCoordinate2D, bounds: MKMapRect) {
+        self.model = MinecraftMapDrawing(drawing: drawing, location: location, mapRect: bounds)
+    }
+
+    public var content: any MinecraftMapContent {
+        MinecraftDrawingOverlay(model: model)
+    }
+}
+
 public class MinecraftDrawingOverlay: NSObject, MKOverlay {
     public var model: MinecraftMapDrawing {
         didSet {
@@ -24,4 +40,8 @@ public class MinecraftDrawingOverlay: NSObject, MKOverlay {
         self.coordinate = model.location
         self.boundingMapRect = model.mapRect
     }
+}
+
+extension MinecraftDrawingOverlay: MinecraftMapContent {
+    public var contentType: MinecraftMapContentType { .overlay }
 }